requirements.txt
asgiref==3.2.5
beautifulsoup4==4.9.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.11
-e git://github.com/Fantomas42/django-blog-zinnia.git@881101a9d1d455b2fc581d6f4ae0947cdd8126c6#egg=django_blog_zinnia
django-classy-tags==1.0.0
django-cms==3.7.1
django-contrib-comments==1.9.2
django-filer==1.7.0
django-filter==2.2.0
django-formtools==2.2
django-haystack==2.8.1
django-js-asset==1.2.2
django-menu==0.1.13
django-menus==1.1.2
django-menus-py3==1.1.3
django-mptt==0.11.0
django-polymorphic==2.1.2
django-sekizai==1.1.0
django-simpleui==3.9.1
django-tagging==0.5.0
django-tagging-ext==0.3.6
django-taggit==1.2.0
django-treebeard==4.3.1
django-xmlrpc==0.1.8
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.2.0
djangocms-bootstrap4==1.5.0
djangocms-file==2.4.0
djangocms-googlemap==1.3.0
djangocms-icon==1.5.0
djangocms-installer==1.2.1
djangocms-link==2.5.0
djangocms-picture==2.3.0
djangocms-snippet==2.2.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.8.0
djangocms-video==2.2.0
djangorestframework==3.10.3
drf-haystack==1.8.6
easy-thumbnails==2.7
gunicorn==20.0.4
html5lib==1.0.1
jieba==0.42.1
Markdown==3.2.1
mots-vides==2015.5.11
Pillow==7.0.0
psycopg2==2.8.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
regex==2020.4.4
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
tzlocal==2.0.0
Unidecode==1.1.1
webencodings==0.5.1
Whoosh==2.7.4
wordpress2zinnia==1.1
static3==0.7.0
whitenoise
simpleui